pyproject.toml
[project]
name = "tricky-mesh-ai-api"
version = "0.1.0"
description = "MeshCore DM → llama.cpp bridge daemon"
requires-python = ">=3.13"
dependencies = [
    "meshcore>=2.3",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]

[project.scripts]
tricky-mesh-ai = "tricky_mesh_ai.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tricky_mesh_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
